% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_scripts.R
\name{show_heellife_examples}
\alias{show_heellife_examples}
\title{Show Example Scripts Information}
\usage{
show_heellife_examples()
}
\description{
Displays information about the available example scripts and how to use them.
}
\examples{
# Show information about example scripts
show_heellife_examples()
}
